Fp16 (#8)
* support fp16 inference for transformer

* more fixes for dtype

* stage code

* stash transformer related changes

* Revert "stash transformer related changes"

This reverts commit 9ca65c55d934bdbdeda766caee8424dd5fde9778.

* revert transformer related chagnes

* cleanup

* fix layernorm dtype

* mp adamw

+++ src/gluonnlp/model/transformer.py
@@ -417,7 +417,7 @@ class BaseTransformerEncoder(HybridBlock, Seq2SeqEncoder):
         length = inputs.shape[1]
         if valid_length is not None:
             mask = mx.nd.broadcast_lesser(
-                mx.nd.arange(length, ctx=valid_length.context).reshape((1, -1)),
+                mx.nd.arange(length, ctx=valid_length.context, dtype=valid_length.dtype).reshape((1, -1)),
                 valid_length.reshape((-1, 1)))
             mask = mx.nd.broadcast_axes(mx.nd.expand_dims(mask, axis=1), axis=1, size=length)
             if states is None:
@@ -889,7 +889,7 @@ class TransformerDecoder(HybridBlock, Seq2SeqDecoder):
         mem_length = mem_value.shape[1]
         if encoder_valid_length is not None:
             mem_masks = mx.nd.broadcast_lesser(
-                mx.nd.arange(mem_length, ctx=encoder_valid_length.context).reshape((1, -1)),
+                mx.nd.arange(mem_length, ctx=encoder_valid_length.context, dtype=encoder_valid_length.dtype).reshape((1, -1)),
                 encoder_valid_length.reshape((-1, 1)))
             decoder_states.append(mem_masks)
         self._encoder_valid_length = encoder_valid_length
@@ -922,13 +922,13 @@ class TransformerDecoder(HybridBlock, Seq2SeqDecoder):
         """
         batch_size = inputs.shape[0]
         length = inputs.shape[1]
-        length_array = mx.nd.arange(length, ctx=inputs.context)
+        length_array = mx.nd.arange(length, ctx=inputs.context, dtype=inputs.dtype)
         mask = mx.nd.broadcast_lesser_equal(
             length_array.reshape((1, -1)),
             length_array.reshape((-1, 1)))
         if valid_length is not None:
             batch_mask = mx.nd.broadcast_lesser(
-                mx.nd.arange(length, ctx=valid_length.context).reshape((1, -1)),
+                mx.nd.arange(length, ctx=valid_length.context, dtype=valid_length.dtype).reshape((1, -1)),
                 valid_length.reshape((-1, 1)))
             mask = mx.nd.broadcast_mul(mx.nd.expand_dims(batch_mask, -1),
                                        mx.nd.expand_dims(mask, 0))
@@ -999,7 +999,7 @@ class TransformerDecoder(HybridBlock, Seq2SeqDecoder):
                 .broadcast_axes(axis=1, size=step_input.shape[1])
             states[-1] = augmented_mem_mask
         if mask is None:
-            length_array = mx.nd.arange(step_input.shape[1], ctx=step_input.context)
+            length_array = mx.nd.arange(step_input.shape[1], ctx=step_input.context, dtype=step_input.dtype)
             mask = mx.nd.broadcast_lesser_equal(
                 length_array.reshape((1, -1)),
                 length_array.reshape((-1, 1)))
